codegen: Большое количество исправлений. Пузырьковая сортировка работает

+++ lexer/lexer.flex.cpp
@@ -2829,7 +2829,7 @@ YY_RULE_SETUP
 
 
     removeUnderline(intLiteral);
-    intValue = atoi(intLiteral); lexer_print("Found a integer value: %d\n", intValue);
+    intValue = atoll(intLiteral); lexer_print("Found a integer value: %d\n", intValue);
     IntType t = IntType::NO_SPEC;
     if (endsWith(lexeme, "S")){
         t = IntType::SHORT_MOD;
